+++ test/core/test_plugin.py
@@ -25,7 +25,7 @@ def test_plugin_compute_request_minimal(
         params=expected_compute_input,
     )
 
-    assert len(computed_artifacts) == 5
+    assert len(computed_artifacts) == 6
     for artifact in computed_artifacts:
         assert isinstance(artifact, Artifact)
 
@@ -52,6 +52,6 @@ def test_plugin_compute_request_all_optionals(
         params=expected_compute_input,
     )
 
-    assert len(computed_artifacts) == 11
+    assert len(computed_artifacts) == 12
     for artifact in computed_artifacts:
         assert isinstance(artifact, Artifact)